+++ tests/transformations/vector_inference_test.py
